spake2: tighten the curve25519-dalek dependency to 1.2.x
This was just "1", which allowed our MSRV to be accidentally raised from 1.31
to 1.32 without a deliberate spake2 minor-version bump (dalek-1.0.x compiles
with rustc-1.31, but dalek-1.1.x required 1.32).

Hopefully by making it "1.2", our MSRV will remain at 1.32 until we
explicitly decide to take on a dependency that needs something newer.

refs #21