travis: use rust 'stable' for rustfmt check
This was previously locked down at a specific version (1.31.0, much older
than the current 1.36), so we'd get consistent behavior over time. But since
we allow the rustfmt build to fail (it's in allow_failures), we don't need
this consistency so much. PR authors are expected to format their code
against the current stable rustfmt, but test it against the MSRV (currently
1.32.0).

refs #21