spake2: switch to latest rand_core and curve25519-dalek
This reduces our dependency load: the latest rand_core includes a basic
OsRng, which is all we need, so we don't need to pull in the entire rand
hierarchy.

Since the curve25519-dalek API includes the RngCore and CryptoRng traits, we
must update to the latest version so they'll match. This also pulls in
zeroize, which uses a feature named `alloc` that did not stabilize until
rustc 1.36.

This increases the 'spake2' MSRV to rust-1.36.